HGemini25ProNode/requirements.txt
# requirements.txt
google-generativeai
Pillow